% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.utils.R
\name{AG.temporal}
\alias{AG.temporal}
\title{AG.temporal}
\usage{
AG.temporal(
  data,
  season = TRUE,
  weekday = FALSE,
  time = FALSE,
  time.breaks = c(0, 9, 15, 19, 24),
  time.labels = c("12-9 am", "9 am-3 pm", "3-7 pm", "After 7 pm")
)
}
\arguments{
\item{data}{Accelerometer data set with a time.stamp variable.}

\item{season}{Boolean value that will return winter, spring, summer, or fall if TRUE, Default: TRUE}

\item{weekday}{Boolean value that will return the day of the week from time.stamp if TRUE, Default: FALSE}

\item{time}{Boolean value that will group the time.stamp into specified categories if TRUE, Default: FALSE}

\item{time.breaks}{Integer values using the 24 hour clock, Default: c(0, 9, 15, 19, 24)}

\item{time.labels}{Labels to be returned from the time.breaks (must be one less value than time.breaks), Default: c("12-9 am", "9 am-3 pm", "3-7 pm", "After 7 pm")}
}
\value{
Returns a data set with the added weekday, time, and season values.
}
\description{
Creates temporal variables from the timestamp that can include season, weekday, and time of the day.
}
\details{
Creates temporal variables from the timestamp that can include season, weekday, and time of the day.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lubridate]{day}}
\code{\link[tibble]{add_column}}
}
