% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_to_csv.R
\name{agd_to_csv}
\alias{agd_to_csv}
\title{Convert AGD to CSV Files}
\usage{
agd_to_csv(data_directory)
}
\arguments{
\item{data_directory}{Location of the folder where the ActiGraph}
}
\value{
Returns a CSV folder in the data directory with the ActiGraph CSV Files
}
\description{
Converts ActiGraph AGD to CSV Files
}
\details{
Converts ActiGraph AGD to CSV Files
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readr]{write_delim}}
}
