% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.utils.R
\name{AGread.csv}
\alias{AGread.csv}
\title{AGread.csv}
\usage{
AGread.csv(demo, newdatadir, file, record_id)
}
\arguments{
\item{demo}{Date of birth file that will be matched based on recrod id to apply age-specific cutpoints.}

\item{file}{Location of the file to be processed.}

\item{record_id}{Participant ID.}
}
\value{
Returns a data set with the timepoint (e.g., A=Baseline, B=6-months, etc.), record ID, timestamp, age of the participant (if date of birth file is used), vertical axis counts, and vector magnitude.
}
\description{
Reads in the accelerometer csv file and prepares the data set for processing.
}
\details{
Reads in the accelerometer csv file and prepares the data set for processing.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readr]{read_delim}}
}
