% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhelselR-package.R
\docType{package}
\name{bhelselR-package}
\alias{bhelselR-package}
\title{bhelselR: Functions for physical activity epidemiology and weight management research}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions for physical activity epidemiology and weight management research.
}
\keyword{internal}
