% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhelselR.R
\docType{package}
\name{bhelselR}
\alias{bhelselR}
\title{bhelselR: A package for computating the notorious bar statistic}
\description{
The bhelselR package provides several important functions:
}
\section{Foo functions}{

agd_to_csv
child_bmi_percentile
dummy_2_categorical
freqper
gt3x2csv
meansd
percent_over_bmi_50
read_agd
}

