% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.utils.R
\name{birth.date}
\alias{birth.date}
\title{birth.date}
\usage{
birth.date(datadir, files)
}
\arguments{
\item{datadir}{Data directory for the accelerometer files.}
}
\value{
Returns a date of birth data set that can be applied to the accelerometer data for age-specific cutpoints.
}
\description{
Reads in a date of birth csv file located in the same data folder as the accelerometer files to apply age-specific cutpoints.
}
\details{
Reads in a date of birth csv file located in the same data folder as the accelerometer files to apply age-specific cutpoints.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readr]{read_delim}}
\code{\link[lubridate]{parse_date_time}}
}
