% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainPower.R
\name{brainpower_add_labels}
\alias{brainpower_add_labels}
\title{brainpower_add_labels}
\usage{
brainpower_add_labels(data, dictionary, add_factor_labels = TRUE)
}
\arguments{
\item{data}{The data set containing the columns and factors to label}

\item{dictionary}{A path to the data dictionary or the data dictionary loaded into R as a dataframe}

\item{add_factor_labels}{Adds labels to factor variables as represented in the data dictionary, Default: TRUE}
}
\value{
A labeled data set
}
\description{
Adds labels and factor labels from the Brain Power data dictionary.
}
\details{
Adds labels and factor labels from the Brain Power data dictionary. The
data dictionary can be accessed after installing the bhelselR package with
\code{system.file("extdata/BrainPowerDataDictionary.csv", package = "bhelselR")}
}
