% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainPower.R
\name{brainpower_merge_files}
\alias{brainpower_merge_files}
\title{brainpower_merge_files}
\usage{
brainpower_merge_files(datadir)
}
\arguments{
\item{datadir}{Data directory where the data files exisit}
}
\value{
The merged data set
}
\description{
Merges the data files from the Brain Power study uploaded to the INCLUDE DCC
}
\details{
Merges the data files from the Brain Power study uploaded to the INCLUDE DCC
}
