% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{calculate_salary}
\alias{calculate_salary}
\title{calculate_salary}
\usage{
calculate_salary(salary, fringe, calmo = NULL, percent = NULL, print = FALSE)
}
\arguments{
\item{salary}{Base salary of key personnel}

\item{fringe}{Fringe rate at the University. The fringe rate can be a decimal
ranging from 0 to < 1 or a numeric variable ranging from 1 to 100.}

\item{calmo}{The effort described in calender year months, Default: NULL}

\item{percent}{The effort described in percent. The percent effort can be a decimal
ranging from 0 to < 1 or a numeric variable ranging from 1 to 100, Default: NULL}

\item{print}{Prints the effort described in calendar months and percent and the
salary, fringe, and total amounts requested, Default: FALSE}
}
\value{
A list of the effort in calendar months and percent and the salary, fringe
and total amount requested.
}
\description{
Input the base salary, fringe rate, and either the effort described
in calendar months or percentage to calculate the salary, fringe, and
total amount requested for key personnel.
}
\details{
Input the base salary, fringe rate, and either the effort described
in calendar months or percentage to calculate the salary, fringe, and
total amount requested for key personnel.
}
\examples{
\dontrun{
if(interactive()){
 calculate_salary(salary = 75000, fringe = 34, percent = 15, print = TRUE)
 }
}
}
