% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIpercentiles.R
\name{child_bmi_percentile}
\alias{child_bmi_percentile}
\title{BMIpercentile}
\usage{
child_bmi_percentile(id, gender, age, bmi)
}
\arguments{
\item{id}{A unique study participant identifier}

\item{gender}{Gender as a character variable, i.e., Male, Female}

\item{age}{Age in years as a numeric variable}

\item{bmi}{Body mass index in kg/m^2 as a numeric variable}
}
\value{
Dataframe with id and body mass index percentile rounded to 4 digits.
}
\description{
Calculate child body mass index percentiles.
}
\details{
This function can be used to calculate child body mass index percentiles for those ≤ 20 years.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
