% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{cor_matrix}
\alias{cor_matrix}
\title{cor_matrix}
\usage{
cor_matrix(
  dataset,
  vars,
  var.names = NULL,
  type = "spearman",
  p.value = 0.05,
  colors = c("#B2BDED", "#C6E0B4")
)
}
\arguments{
\item{dataset}{The data set where the variables can be located.}

\item{vars}{The names of the variables as they appear in the data set.}

\item{var.names}{The names of the variables as they should appear on the plot. If no variable names are entered, the plot will use the variable names from the data set, Default: NULL}

\item{type}{Type of correlation (i.e., spearman or pearson), Default: 'spearman'}

\item{p.value}{The p-value threshold to use for significance. A table of p-values under this threshold will be returned as the second element of this list, Default: 0.05}

\item{colors}{The low and high end colors on the plot to represent strong negative and positive correlations, respectively, Default: c("#B2BDED", "#C6E0B4")}
}
\value{
Returns a list where the first element is the plot and the second element is a table of p-values that are under the p.value threshold
}
\description{
A function to visualize correlations between variables.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Hmisc]{rcorr}}
\code{\link[reshape2]{melt}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}
}
