% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.cutpoints.R
\name{cutpoints}
\alias{cutpoints}
\title{cutpoints}
\usage{
cutpoints(data, sets, set.name, n.axis, spurious = 20000)
}
\arguments{
\item{data}{Accelerometer data set with a time stamp and either counts or vector magnitude.}

\item{set.name}{Cutpoint set to be used. Options currently include freedson.child, troiano.adult, fariasbarnett.adult, or freedson.adult.}

\item{n.axis}{Number of axes to be used. Options currently include 1 (vertical axis) or 3 (vector magnitude)}
}
\value{
Returns accelerometer data set with an intensity value that also include nonwear time.
}
\description{
This function applies commonly used cutpoints to accelerometer data and returns the intensity.
}
\details{
This function applies commonly used cutpoints to accelerometer data and returns the intensity.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DescTools]{Mode}}
\code{\link[stringr]{str_split}}
\code{\link[PhysicalActivity]{wearingMarking}}
}
