% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.cutpoints.R
\name{detect.bouts}
\alias{detect.bouts}
\title{Detect bouts}
\usage{
detect.bouts(data, bout = 10, tolerance = 2)
}
\arguments{
\item{data}{Data that has MVPA has a binary variable with 1 indicating MVPA and a days variable (integer) to indicate study day}

\item{bout}{The threshold to consider as a bout of physical activity, Default: 10}

\item{tolerance}{Minute allowance greater than the tolerance that is required to terminate a bout of physical activity, Default: 2}
}
\value{
Dataset with the added bout length and bout count variables
}
\description{
Detect bouts of moderate-to-vigorous physical activity
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
