% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{dummy_2_categorical}
\alias{dummy_2_categorical}
\title{Dummy 2 Categorical}
\usage{
dummy_2_categorical(data, id, new_var, column_list, new_names)
}
\arguments{
\item{data}{Data set where the variables can be located.}

\item{id}{The unique identifier for the participants in the data.}

\item{new_var}{The name of the new categorical variable that will be created.}

\item{column_list}{Combined list using c() of the data columns containing the dummy variables.}

\item{new_names}{The new categories that will be created within the categorical variable.}
}
\value{
Returns the data set with the new categorical variable.
}
\description{
Converts several related dummy variables to a single categorical variable.
}
\details{
Converts several related dummy variables to a single categorical variable.
}
\examples{
\dontrun{
 #Add examples
}
}
