% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{extract_reference_table}
\alias{extract_reference_table}
\title{extract_reference_table}
\usage{
extract_reference_table(path, years = NULL, return_data_frame = FALSE)
}
\arguments{
\item{path}{The path name for the HTML page containing a curriculum vita downloaded from the faculty member's profile}

\item{years}{The year (e.g., 2024) or years (e.g., 2022:2024 or c(2022, 2023, 2024))
of which to extract the references. If the years argument is left blank, the
function retrieves references for the current year, Default: NULL}

\item{return_data_frame}{Combine all the journal articles that have been published,
in press, or submitted into a single data frame, Default: FALSE}
}
\value{
A list or data frame containing the references for the designated years
}
\description{
Extracts the reference table from a faculty curriculum vitae
containing journal articles that have been published, in press, or submitted
}
\details{
Extracts the reference table from a faculty curriculum vitae
containing journal articles that have been published, in press, or submitted
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{bind_rows}}
\code{\link[stringr]{str_extract}}, \code{\link[stringr]{str_remove}}
\code{\link[purrr]{map2}}
}
