% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{find_table}
\alias{find_table}
\title{find_table}
\usage{
find_table(path, tag, label)
}
\arguments{
\item{path}{The path name for the HTML page containing a curriculum vita downloaded from the faculty member's profile}

\item{tag}{The HTML tag containing label}

\item{label}{Text that labels the start of the HTML table}
}
\value{
A tibble containing the table from the HTML page
}
\description{
Reads in an HTML page from a faculty curriculum vitae and extracts the table following the HTML tag and label
}
\details{
Reads in an HTML page from a faculty curriculum vitae and extracts the table following the HTML tag and label
}
\seealso{
\code{\link[rvest]{read_html}}, \code{\link[rvest]{html_element}}, \code{\link[rvest]{html_table}}
}
