% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{freqper}
\alias{freqper}
\title{Frequency and Percentage}
\usage{
freqper(x, total)
}
\arguments{
\item{x}{Name of a categorical or binary variable in data.}

\item{total}{Number of total observations for the categorical or binary variable.}
}
\value{
Frequency and percentage of the categorical or binary variable.
}
\description{
Prints out the frequency and percentage for a categorical or binary variable.
}
\details{
Prints out the frequency and percentage for a categorical or binary variable.
}
\examples{
\dontrun{
#EXAMPLE1
x = rnorm(100, 40, 5)
y = ifelse(x < 40, 1, 0)
freqper(y, 100)
 }
}
