% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readActigraph.R
\name{gt3x2csv}
\alias{gt3x2csv}
\title{Convert GT3X Files to CSV}
\usage{
gt3x2csv(origin, output, studyname)
}
\arguments{
\item{origin}{Origin of the GT3X file.}

\item{output}{Desired location of the CSV file.}

\item{studyname}{Study name or abbreviation to be used in the naming of the file.}
}
\value{
Returns a CSV file to the path name specified under output.
}
\description{
Convert a single GT3X file to a CSV file with removed timestamp to be processed by GGIR.
}
\details{
Convert a single GT3X file to a CSV file with removed timestamp to be processed by GGIR.
}
\examples{
\dontrun{
 #Add example here.
}
}
