% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{ilr_transform}
\alias{ilr_transform}
\title{Isometric Log Ratio Transformation}
\usage{
ilr_transform(X)
}
\arguments{
\item{X}{Matrix or data frame that includes only variables in the composition.}
}
\value{
A matrix with (n-1) columns containing the isometric log ratio
transformed values.
}
\description{
This function does an isometric log ratio transformation.
}
\details{
This function does an isometric log ratio transformation as
described by \href{https://pubmed.ncbi.nlm.nih.gov/26461112/}{Chastin et al., 2005}.
}
