% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.main}
\alias{mars.main}
\title{Movement Analysis for Remote Sensors (MARS)}
\usage{
mars.main(
  study.name = "study",
  study.id = "A1001",
  datadir,
  results,
  adult.cp = "",
  child.cp = "",
  axis = 1,
  overwrite = FALSE,
  person.time = FALSE,
  person.date = TRUE,
  person.month = FALSE,
  valid = 480,
  return.timestamped.dataframe = FALSE,
  individual.file.save = FALSE
)
}
\arguments{
\item{study.name}{Abbreviation of the Study, Default: study}

\item{study.id}{Example of a study id number with time point (A: Baseline, B: 6 months, etc.), Default: 'A1001'}

\item{datadir}{Directory where the AGD files are stored.}

\item{results}{Directory where the results should be stored.}

\item{adult.cp}{Adult cut-point to be used (see cutpoint.list.R for list of available cut-points), Default: ''}

\item{child.cp}{Child cut-point to be used (see cutpoint.list.R for list of available cut-points), Default: ''}

\item{axis}{Number of axes to be used (1: Vertical Axis; 3: Vector Magnitude), Default: 1}

\item{overwrite}{Overwrite the individual files if individual.file.save is set to TRUE, Default: FALSE}

\item{person.time}{Export the accelerometer summary by person time categories, Default: FALSE}

\item{person.date}{Export the accelerometer summary by person date, Default: TRUE}

\item{person.month}{Export only valid days to a summary by person time point of the trial, Default: FALSE}

\item{valid}{Set the number of minutes for a valid day, Default: 480}

\item{return.timestamped.dataframe}{Return a timestamped data set into R for all combined time point and id data, Default: FALSE}

\item{individual.file.save}{Save all of the individual timestamped files to a CSV, Default: FALSE}
}
\value{
Individual and/or summary files to a Results folder located in the designated directory.
}
\description{
Main function to execute MARS accelerometer processing program.
}
\details{
Main function to execute MARS accelerometer processing program.
}
\examples{
\dontrun{
if(interactive()){
 mars.main(study.name = "study", study.id = "A1001", datadir=datadir, 
 results=results, adult.cp = "troiano.adult", child.cp = "freedson.child", 
 axis = 1, overwrite=FALSE, person.time=FALSE,  person.date=TRUE, 
 person.month=FALSE, valid = 480, return.timestamped.dataframe=FALSE, 
 individual.file.save=FALSE)
 }
}
}
\seealso{
\code{\link[readr]{write_delim}}
\code{\link[plyr]{rbind.fill}}
}
