% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{meansd}
\alias{meansd}
\title{Mean and Standard Deviation}
\usage{
meansd(x)
}
\arguments{
\item{x}{Name of a continuous variable in data.}
}
\value{
Mean and Standard Deviation.
}
\description{
Prints out mean and standard deviation for a continuous variable.
}
\details{
Prints out mean and standard deviation for a continuous variable.
}
\examples{
\dontrun{
 x = rnorm(100, 40, 5)
 meansd(x)
 }
}
