% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvoREE.R
\name{parvo.ree.extract.data}
\alias{parvo.ree.extract.data}
\title{Parvo Extract Data}
\usage{
parvo.ree.extract.data(parvo.path)
}
\arguments{
\item{parvo.path}{Pathname to the Parvo XLSX file.}
}
\value{
Returns a data set with the Parvo data.
}
\description{
Extracts the Parvo data from the XLSX file.
}
\details{
Extracts the Parvo data from the XLSX file.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[lubridate]{as_date}},\code{\link[lubridate]{round_date}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{character(0)}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}}
}
