% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvoREE.R
\name{parvo.ree.main}
\alias{parvo.ree.main}
\title{Resting Energy Expenditure Main Function}
\usage{
parvo.ree.main(accel.path = NULL, parvo.path, var.limit = 15)
}
\arguments{
\item{accel.path}{Pathname to the accelerometer AGD file, Default: NULL.}

\item{parvo.path}{Pathname to the Parvo XLSX file.}

\item{var.limit}{Number to ensure changes in ventilation, oxygen consumption, and respiratory quotient do not exceed variation limit.}
}
\value{
Returns a matrix with the resting energy expenditure output, the length of time of steady state, and the number of the last observations used (maximum: 5).
}
\description{
Read in Parvo and Accelerometer data (if available) and calculate the estimated resting energy expenditure.
}
\details{
Read in Parvo and Accelerometer data (if available) and calculate the estimated resting energy expenditure.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lubridate]{period}},\code{\link[lubridate]{round_date}}
\code{\link[dplyr]{lead-lag}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{rename}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{context}},\code{\link[dplyr]{select}}
\code{\link[bhelselR]{read_agd}}
}
