% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_oues.R
\name{parvo_get_oues}
\alias{parvo_get_oues}
\title{parvo_get_oues}
\usage{
parvo_get_oues(
  path,
  return.data = TRUE,
  write.summary.file = FALSE,
  write.individual.file = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{path}{Path to a single file or directory with the VO2 Max excel files}

\item{return.data}{Returns the OUES results to the R console, Default: TRUE}

\item{write.summary.file}{Exports a summary CSV file of VO2 max characteristics for study participants, Default: FALSE}

\item{write.individual.file}{Exports individual CSV files for study participants, Default: FALSE}

\item{verbose}{Print the progress of the OUES calculation, Default: FALSE}

\item{...}{arguments passed to \code{\link[data.table]{fwrite}}}
}
\value{
Returns the OUES results or writes individual or summary files
}
\description{
A function to calculate Oxygen Uptake Efficiency Slope at 75\%, 90\%, and 100\% VO2 max measured using a Parvo Medics TrueOne 2400 metabolic cart
}
\details{
A function to calculate Oxygen Uptake Efficiency Slope at 75\%, 90\%, and 100\% VO2 max measured using a Parvo Medics TrueOne 2400 metabolic cart
}
\examples{
\dontrun{
if(interactive()){
 parvo_get_oues(path)
 }
}
}
\seealso{
\code{\link[stringr]{str_split}}
\code{\link[readxl]{read_excel}}
\code{\link[data.table]{fwrite}}
\code{\link[plyr]{rbind.fill}}
\code{\link[stats]{lm}}
}
