% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readActigraph.R
\name{read_agd}
\alias{read_agd}
\title{Read ActiGraph AGD Files}
\usage{
read_agd(agd_filename, settings = FALSE, data = TRUE)
}
\arguments{
\item{agd_filename}{The path to the AGD file or the filename if the AGD file is in the current working directory.}

\item{settings}{Returns the AGD file meta data that is reported in ActiGraph CSV Files in a data frame, Default: FALSE}

\item{data}{Returns the AGD file accelerometer data in a data frame, Default: TRUE}
}
\value{
Returns a data frame or a list if settings and data are both TRUE
}
\description{
Read the settings or data from the ActiGraph AGD Files
}
\details{
Read the settings or data from the ActiGraph AGD Files
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DBI]{dbConnect}},\code{\link[DBI]{dbReadTable}},\code{\link[DBI]{dbDisconnect}}
\code{\link[RSQLite]{SQLite}}
}
