% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{retrieve_redcap_report}
\alias{retrieve_redcap_report}
\title{retrieve_redcap_report}
\usage{
retrieve_redcap_report(study, report_id)
}
\arguments{
\item{study}{A study acronym or name to select the token from a user's R environment (stored as STUDY_TOKEN)}

\item{report_id}{A numeric value representing the REDCap report to be downloaded to the user's R environment}
}
\value{
A data frame containing the extracted report
}
\description{
Retrieves a report from the REDCap API
}
\details{
Retrieves a report from the REDCap API
}
\seealso{
\code{\link[RCurl]{postForm}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
