% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_study_week.R
\name{retrieve_study_week}
\alias{retrieve_study_week}
\title{retrieve_study_week}
\usage{
retrieve_study_week(data, date, first_day = "Monday", id)
}
\arguments{
\item{data}{The data set with at least a date and id column}

\item{date}{The name of the column storing the sequential dates}

\item{first_day}{Day of the week that the study starts, Default: 'Monday'}

\item{id}{The name of the column storing the participant IDs}
}
\value{
A tibble with the study weeks added to the data set
}
\description{
Retrieves the study week for each person in a data set with sequential dates
}
\details{
Retrieves the study week for each person in a data set with sequential dates. Zeros are
added if the dates start before first_day.
}
\seealso{
\code{\link[dplyr]{group_by_all}}, \code{\link[dplyr]{summarise}}
\code{\link[lubridate]{day}}
}
