% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendmail.R
\name{sendmail}
\alias{sendmail}
\title{sendmail}
\usage{
sendmail(
  host,
  port,
  sender,
  recipients,
  subject,
  body,
  attachments = NULL,
  html = FALSE
)
}
\arguments{
\item{host}{The name of the remote host to which to connect}

\item{port}{The specific port of the remote host to which to connect}

\item{sender}{The email address of the senders. The sender's email can either be just an email or a combined vector of name and email.}

\item{recipients}{The email address of the recipients either as a single email or a combined vector of emails}

\item{subject}{The subject of the email}

\item{body}{The body of the email either in HTML or plain text}

\item{attachments}{Full path name to single file or a directory of files that should be included in the email, Default: NULL}

\item{html}{Set this parameter to TRUE if the body contains HTML, Default: FALSE}
}
\value{
Sends an email via R
}
\description{
Sends an email in R using the reticulate package as an R-Python interface for the Python email package
}
\details{
Sends an email in R using the reticulate package as an R-Python interface for the Python email package
}
\seealso{
\code{\link[reticulate]{import}}, \code{\link[reticulate]{r-py-conversion}}
}
