% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_downloads.R
\name{sort_files_in_downloads}
\alias{sort_files_in_downloads}
\title{Sort Files in Downloads Folder}
\usage{
sort_files_in_downloads(allowed_filetypes = NULL)
}
\arguments{
\item{allowed_filetypes}{Add file types, Default: NULL}
}
\value{
A sorted downloads folder
}
\description{
This function allows a user to clean up their downloads folder.
}
\details{
This function allows a user to clean up their downloads folder. By
changing the allowed file types, a user can specify additional files that
they want to keep in the downloads folder.
}
\examples{
sort_files_in_downloads(allowed_filetypes = c("zip", "txt", "R"))
}
\seealso{
\code{\link[tools]{fileutils}}
}
