% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhelselR.R
\docType{package}
\name{bhelselR}
\alias{bhelselR}
\title{bhelselR: A package for physical activity and weight management research.}
\description{
The bhelselR package provides several important functions:
}
\section{bhelselR functions}{

\itemize{
\item \code{\link[=readActigraph]{readActigraph()}}
\item child_bmi_percentile
\item dummy_2_categorical
\item freqper
\item gt3x2csv
\item meansd
\item percent_over_bmi_50
\item read_agd
}
}

