% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhelselR.R
\docType{package}
\name{bhelselR}
\alias{bhelselR}
\title{bhelselR: A package for physical activity and weight management research.}
\description{
The bhelselR package provides several important functions:

\code{\link{child_bmi_percentile}}

\code{\link{percent_over_bmi_50}}

\code{\link{dummy_2_categorical}}

\code{\link{freqper}}

\code{\link{meansd}}
}
\section{Visualization}{


\code{\link{cor_matrix}}
}

\section{Utilities}{


\code{\link{sort_files_in_downloads}}

\code{\link{dpawmReliabilityTesting}}
}

