% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhelselR.R
\docType{package}
\name{bhelselR}
\alias{bhelselR}
\title{bhelselR: A package for physical activity and weight management research.}
\description{
The bhelselR package provides several important functions:
}
\section{bhelselR functions}{

\itemize{
\item \link[=child_bmi_percentile]{child_bmi_percentile}
\item \link[=dummy_2_categorical]{dummy_2_categorical}
\item \link[=freqper]{freqper}
\item \link[=gt3x2csv]{gt3x2csv}
\item \link[=meansd]{meansd}
\item \link[=percent_over_bmi_50]{percent_over_bmi_50}
\item \link[=read_agd]{read_agd}
}
}

