% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhelselR.R
\docType{package}
\name{bhelselR}
\alias{bhelselR}
\title{bhelselR: A package for physical activity and weight management research.}
\description{
The bhelselR package provides several important functions:
}
\section{bhelselR functions}{

\code{\link{dummy_2_categorical}}

\code{\link{child_bmi_percentile}}

\code{\link{freqper}}

\code{\link{gt3x2csv}}

\code{\link{meansd}}

\code{\link{percent_over_bmi_50}}

\code{\link{read_agd}}
}

