% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhelselR.R
\docType{package}
\name{bhelselR}
\alias{bhelselR}
\title{bhelselR: A package for physical activity and weight management research.}
\description{
The bhelselR package provides several important functions:
}
\section{Data Processing}{


\code{\link{child_bmi_percentile}}

\code{\link{percent_over_bmi_50}}

\code{\link{dummy_2_categorical}}

\code{\link{freqper}}

\code{\link{meansd}}
}

\section{Accelerometer}{


\code{\link{mars.main}}

\code{\link{read_agd}}

\code{\link{agd_to_csv}}

\code{\link{gt3x2csv}}

\code{\link{AGread.csv}}

\code{\link{birth.date}}

\code{\link{AG.temporal}}

\code{\link{cutpoints}}

\code{\link{detect.bouts}}
}

\section{Energy Expenditure}{


\code{\link{parvo.extract.data}}

\code{\link{parvo.extract.meta}}

\code{\link{parvo.ree.main}}

\code{\link{parvo.aee.final4}}
}

\section{Visualization}{


\code{\link{cor_matrix}}
}

