% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvoREE.R
\name{parvo.aee.final4}
\alias{parvo.aee.final4}
\title{Parvo AEE Final 4}
\usage{
parvo.aee.final4(parvo.path)
}
\arguments{
\item{parvo.path}{Pathname to the Parvo XLSX file.}
}
\value{
Returns an average for the last 4 minutes of the WalkDS walking stages for VO2, METS, and RQ.
}
\description{
Takes average of last 4 minutes of the walking protocol from the WalkDS study.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[lubridate]{as_date}},\code{\link[lubridate]{round_date}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{character(0)}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}}
}
