% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parvoREE.R
\name{parvo.aee.final4}
\alias{parvo.aee.final4}
\title{Parvo AEE Final 4}
\usage{
parvo.aee.final4(
  parvo.path,
  corrected.time.path = NULL,
  accel.path = NULL,
  rest1met = 3.5
)
}
\arguments{
\item{parvo.path}{Pathname to the Parvo XLSX file.}

\item{corrected.time.path}{Pathname to a document of corrected time stamps, Default: NULL.}

\item{accel.path}{Pathname to the accelerometer AGD file, Default: NULL.}

\item{rest1met}{Resting VO2 for 1 metabolic equivalent (MET), Default = 3.5 ml/kg/min}
}
\value{
Returns an average for the last 4 minutes of the WalkDS walking stages for VO2, METS, and RQ.
}
\description{
Takes average of last 4 minutes of the walking protocol from the WalkDS study.
}
\details{
Takes average of last 4 minutes of the walking protocol from the WalkDS study.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[dplyr]{select}}
}
